pulizia codice

+++ frontend/src/components/componentsStaff/BooksStaffCatalog.jsx
@@ -3,6 +3,7 @@ import { deleteBooks, getBooks } from "../../services/bookService";
 import Spinner from "../Spinner";
 import { useNavigate } from "react-router-dom";
 import ConfirmModal from "./ConfirmModal";
+import { useAuth } from "../../hooks/useAuth";
 
 const BooksStaffCatalog = () => {
   const [books, setBooks] = useState([]);
@@ -12,12 +13,12 @@ const BooksStaffCatalog = () => {
   const [search, setSearch] = useState("");
   const [currentPage, setCurrentPage] = useState(0);
   const [totalPages, setTotalPages] = useState(1);
-  const token = localStorage.getItem("authToken")
-    ? localStorage.getItem("authToken")
-    : sessionStorage.getItem("authToken");
   const [modalOpen, setModalOpen] = useState(false);
   const [bookToDelete, setBookToDelete] = useState(null);
   const navigate = useNavigate();
+  const { user } = useAuth();
+  const token =
+    localStorage.getItem("authToken") || sessionStorage.getItem("authToken");
 
   const loadBooks = async (page = currentPage) => {
     setLoading(true);
@@ -51,18 +52,26 @@ const BooksStaffCatalog = () => {
 
   const handleConfirmDelete = async () => {
     if (!bookToDelete) return;
+
+    if (user.role === "DEMO") {
+      alert("Gli utenti demo non possono eliminare libri.");
+      setModalOpen(false);
+      setBookToDelete(null);
+      return;
+    }
+
     try {
       setLoading(true);
       setModalOpen(false);
       await deleteBooks(bookToDelete, token);
       await loadBooks();
+      alert("Libro eliminato con successo");
     } catch (err) {
       setError(err);
+      alert(`Errore durante l'eliminazione: ${err.message}`);
     } finally {
       setLoading(false);
-
       setBookToDelete(null);
-      alert("Libro eliminato con successo");
     }
   };
 
@@ -120,18 +129,18 @@ const BooksStaffCatalog = () => {
 
   return (
     <div className="w-full space-y-4">
-      {/* Intestazione + Aggiungi Libro */}
       <div className="flex justify-between items-center mb-4">
         <h1 className="text-2xl font-bold">Gestione Libri</h1>
-        <button
-          className="px-4 py-2 bg-blue-500 text-white font-semibold hover:bg-blue-600 transition"
-          onClick={() => navigate("/admin/add-book")}
-        >
-          Aggiungi Nuovo Libro
-        </button>
+        {(user.role === "STAFF" || user.role === "DEMO") && (
+          <button
+            className="px-4 py-2 bg-blue-500 text-white font-semibold hover:bg-blue-600 transition"
+            onClick={() => navigate("/admin/add-book")}
+          >
+            Aggiungi Nuovo Libro
+          </button>
+        )}
       </div>
 
-      {/* Barra di ricerca */}
       <input
         className="w-full p-3 bg-gray-100 border border-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-400 placeholder-gray-500"
         type="text"
@@ -141,7 +150,6 @@ const BooksStaffCatalog = () => {
         onChange={(e) => setSearch(e.target.value)}
       />
 
-      {/* Intestazione tabella */}
       <div className="grid grid-cols-6 gap-4 bg-gray-100 p-3 font-semibold border border-gray-300">
         <p>Titolo</p>
         <p>Autore</p>
@@ -151,7 +159,6 @@ const BooksStaffCatalog = () => {
         <p>Azioni</p>
       </div>
 
-      {/* Contenuto tabella */}
       {loading ? (
         <Spinner />
       ) : error ? (
@@ -176,25 +183,28 @@ const BooksStaffCatalog = () => {
             </p>
             <p>{book.position || "-"}</p>
             <div className="flex gap-2">
-              <button
-                className="text-blue-500 hover:text-blue-700 font-medium transition"
-                onClick={() => navigate(`/admin/update-book/${book.id}`)}
-              >
-                Modifica
-              </button>
-              <span className="text-gray-400">|</span>
-              <button
-                className="text-red-500 hover:text-red-700 font-medium transition"
-                onClick={() => handleDeleteClick(book.id)}
-              >
-                Elimina
-              </button>
+              {(user.role === "STAFF" || user.role === "DEMO") && (
+                <>
+                  <button
+                    className="text-blue-500 hover:text-blue-700 font-medium transition"
+                    onClick={() => navigate(`/admin/update-book/${book.id}`)}
+                  >
+                    Modifica
+                  </button>
+                  <span className="text-gray-400">|</span>
+                  <button
+                    className="text-red-500 hover:text-red-700 font-medium transition"
+                    onClick={() => handleDeleteClick(book.id)}
+                  >
+                    Elimina
+                  </button>
+                </>
+              )}
             </div>
           </div>
         ))
       )}
 
-      {/* Paginazione */}
       {renderPagination()}
 
       <ConfirmModal
